Index-About Page
Completed work on index/home page and about us... The same navbar is placed at the top and footer at the bottom of every page to keep the design consistent. Navbar links connect all pages, and the active page link is highlighted. This makes the website look professional and easy to navigate.